requirements.txt
gtts==2.2.3
joblib==1.0.1
pillow==8.3.1
pip-chill==1.0.1
playsound==1.2.2
pyaudio==0.2.11
pyttsx3==2.90
regex==2021.7.6
speech-recognition-fork==3.8.1.2021.6.14
tqdm==4.61.2
